login logic refactoring
