+++ pins/ViewController/LoginViewController.swift
@@ -8,12 +8,11 @@
 import UIKit
 import OSLog
 import AuthenticationServices
-import CryptoKit
 import FirebaseAuth
 
 class LoginViewController: UIViewController {
-    private var currentNonce: String?
-
+    let viewModel: LoginViewModel = LoginViewModel()
+    
     var loginView: LoginView {
         view as! LoginView
     }
@@ -29,51 +28,9 @@ class LoginViewController: UIViewController {
     
     private func setAction() {
         loginView.setLoginAction(UIAction(handler: { [weak self] _ in
-            self?.startSignInWithAppleFlow()
+            guard let self = self else { return }
+            self.viewModel.appleLogin(delegate: self)
         }))
-    }
-    
-    func startSignInWithAppleFlow() {
-        let nonce = randomNonceString()
-        currentNonce = nonce
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
-    }
-    
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            return String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
-    }
-    
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        var randomBytes = [UInt8](repeating: 0, count: length)
-        let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
-        
-        if errorCode != errSecSuccess {
-            fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
-        }
-
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-
-        let nonce = randomBytes.map { byte in
-            charset[Int(byte) % charset.count]
-        }
-
-        return String(nonce)
     }
 }
 
@@ -84,7 +41,7 @@ extension LoginViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
     
     func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
         if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
-            guard let nonce = currentNonce else {
+            guard let nonce = viewModel.getAppleNonce() else {
                 fatalError("Invalid state: A login callback was received, but no login request was sent.")
             }
             guard let appleIDToken = appleIDCredential.identityToken else {
